requirements.txt
alembic==1.13.1
Flask==3.0.3
Flask_Migrate==4.0.7
flask_sqlalchemy==3.1.1
numpy==1.24.4
pandas==2.0.3
psycopg2==2.9.9
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==0.24.1
SQLAlchemy==2.0.30
